+++ webpack/ui-components/src/index.js
@@ -1,3 +1,4 @@
+import "./index.css"
 import {
   Accordion,
   AccordionList,
@@ -136,9 +137,9 @@ import {
 // import { initCoreLibraries } from "@egovernments/digit-ui-libraries-core";
 import { initLibraries } from "@egovernments/digit-ui-libraries";
 
-// initLibraries().then(() => {
-//   console.info("DIGIT Contants enabled");
-// });
+initLibraries().then(() => {
+  console.info("DIGIT Contants enabled");
+});
 
 export {
   Footer,