+++ webpack/ui-components/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0-beta.40-rc19.02] [28-Jun-2025]
+-New css classes introduced in index.css and submitbar.css for keyboard focus accessibilty
+-Keyboard and screen reader accessibilty implemented to each atoms, molecules and HOCs
+
 ## [0.2.0-beta.40-rc19.01] [27-Jun-2025]
 - ui-components(0.2.0-beta.40) from develop Upgraded to react19
 - webpack for build